декомпозиция приложения
